Move logic to a trait
